% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_gee.R
\name{a_lsmeans_logistic}
\alias{a_lsmeans_logistic}
\title{Extract Logistic GEE Model LS Means Estimates}
\usage{
a_lsmeans_logistic(df, .in_ref_col)
}
\description{
Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.
}
