% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_gee.R
\name{as.rtable.tern_gee}
\alias{as.rtable.tern_gee}
\title{Extract the coefficient table or covariance matrix estimate from a \code{tern_gee} object.}
\usage{
\method{as.rtable}{tern_gee}(x, type = c("coef", "cov"), ...)
}
\description{
Extract the coefficient table or covariance matrix estimate from a \code{tern_gee} object.
}
