% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_gee.R
\name{s_lsmeans_logistic}
\alias{s_lsmeans_logistic}
\title{Extract Logistic GEE Model LS Means Estimates}
\usage{
s_lsmeans_logistic(df, .in_ref_col)
}
\arguments{
\item{df}{(\code{data.frame})\cr data set resulting from \code{\link[=lsmeans]{lsmeans()}}.}

\item{.in_ref_col}{(\code{logical})\cr \code{TRUE} when working with the reference level, \code{FALSE} otherwise.}
}
\description{
Statistics function which extracts estimates from a \code{\link[=lsmeans]{lsmeans()}} data frame based on a logistic GEE model.
}
\examples{
df <- fev_data
df$AVAL <- rbinom(n = nrow(df), size = 1, prob = 0.5)
lsmeans_df <- lsmeans(fit_gee(vars = vars_gee(arm = "ARMCD"), data = df))

s_lsmeans_logistic(lsmeans_df[1,], .in_ref_col = TRUE)

s_lsmeans_logistic(lsmeans_df[2,], .in_ref_col = FALSE)
}
