% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_gee.R
\name{summarize_gee_logistic}
\alias{summarize_gee_logistic}
\title{Summarize Logistic GEE Least Square Mean Results}
\usage{
summarize_gee_logistic(
  lyt,
  ...,
  table_names = "lsmeans_logistic_summary",
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)
}
\arguments{
\item{lyt}{(\code{layout})\cr input layout where analyses will be added to.}

\item{table_names}{(\code{character})\cr this can be customized in case that the same \code{vars}
are analyzed multiple times, to avoid warnings from \code{rtables}.}

\item{.stats}{(\code{character})\cr statistics to select for the table.}

\item{.formats}{(named \code{character} or \code{list})\cr formats for the statistics.}

\item{.indent_mods}{(named \code{integer})\cr indent modifiers for the labels.}

\item{.labels}{(named \code{character})\cr labels for the statistics (without indent).}
}
\description{
Analyze function for tabulating least-squares means estimates from logistic GEE least square mean results.
}
\examples{
library(dplyr)

df <- fev_data \%>\%
  mutate(AVAL = rbinom(n = nrow(df), size = 1, prob = 0.5))
  df_all <- df \%>\%
    select(USUBJID, ARMCD) \%>\%
      unique()

lsmeans_df <- lsmeans(fit_gee(vars = vars_gee(arm = "ARMCD"), data = df))
basic_table() \%>\%
  split_cols_by("ARMCD") \%>\%
  add_colcounts() \%>\%
  summarize_gee_logistic(
    .in_ref_col = FALSE
  ) \%>\%
  build_table(lsmeans_df, alt_counts_df = df_all)
}
