% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_methods.R
\name{gee_methods}
\alias{gee_methods}
\alias{VarCorr.tern_gee}
\title{Method for GEE Models}
\usage{
\method{VarCorr}{tern_gee}(x, sigma = 1, ...)
}
\description{
Additional methods which can simplify working with the GEE result object.
}
