% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern.gee-package}
\alias{tern.gee}
\alias{tern.gee-package}
\title{\code{tern.gee} Package}
\description{
Create tables and graphs for GEE model fits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/tern.gee/}
  \item \url{https://github.com/insightsengineering/tern.gee/}
  \item Report bugs at \url{https://github.com/insightsengineering/tern.gee/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})

Authors:
\itemize{
  \item Daniel Sabanés Bové \email{daniel@rconis.com}
  \item Emily de la Rua \email{emily.de_la_rua@contractors.roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
