% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern.gee-package}
\alias{tern.gee}
\alias{tern.gee-package}
\title{\code{tern.gee} Package}
\description{
Create tables and graphs for GEE model fits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/tern.gee/}
  \item Report bugs at \url{https://github.com/insightsengineering/tern.gee/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanés Bové \email{daniel.sabanes_bove@roche.com}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
